lambda_alert_status/requirements.txt
boto3==1.15.1
numpy==1.19.5
pandas==1.1.2
PyMySQL==1.0.2
logger==1.4
PyYAML==6.0
